Move form and projects state into portfolio
